+++ mvlearn/utils/__init__.py
@@ -1,4 +1,3 @@
 from .utils import check_Xs
 from .utils import check_Xs_y
 from .utils import check_Xs_y_nan_allowed
-from .dcca_utils import linear_cca, cca_loss, MlpNet, DeepCCA
